fix: /seeting/userprofile 페이지 없애고 /profile/my 페이지 이동 #88
